Highlight graph series when Legend items are hovered (#55)
This allows the user to highlight individual Line or Bar series in a Chart by hovering over items in the Legend. It works by defining a `ReportCardContext` through the [React Context API](https://reactjs.org/docs/context.html) and sharing it with a Legend and Chart inside a ReportCard. This is a useful way to implement the solution because all of the wiring happens behind the scenes. No new props are added to any of the components and the app developer doesn't have to do anything to make it work.